requirements.txt
pycryptodome==3.21.0
simplejson==3.19.3
pbkdf2==1.3